marius

+++ Projet/server/routes/v1/sheet.js
@@ -1,5 +1,6 @@
 import sheetRepository from "../../persistence/sheet-repository.js";
 import express from 'express';
+import sheetValidator from '../../validators/sheet-validator.js';
 
 const router = express.Router();
 
@@ -14,6 +15,7 @@ router.get('/sheets', async (req, res, next) => {
 
 router.get('/sheets/:id', async (req, res, next) => {
     try {
+
         const id = parseInt(req.params.id);
         const sheet = await sheetRepository.findSheet(id);
         if (sheet) {
@@ -36,16 +38,29 @@ router.post('/sheets', async (req, res, next) => {
         const id_creator = req.body.id_creator
         const imageData = req.body.imageData
 
-        const sheet = await sheetRepository.createSheet(title, group, instrument, difficulty, id_creator,imageData);
-        res.status(201).json(sheet);
-    } catch (error) {
-        next(error);
+        const errors = sheetValidator.validateCreateSheet(title, group, instrument, difficulty, id_creator, imageData);
+        console.log("Effeur :",errors)
+        if (errors.length > 0) {
+            res.status(400).json({errors});
+            return next(errors);
+        }
+        else{
+        const sheet = await sheetRepository.createSheet(title, group, instrument, difficulty, id_creator, imageData);
+        res.status(201).json(sheet);}
+    } catch (errors) {
+        next(errors);
+
     }
 });
 
 router.delete('/sheets/:id', async (req, res, next) => {
     try {
         const id = parseInt(req.params.id);
+        /*const errors = sheetValidator.validateDeleteSheet(id);
+        if (errors.length > 0) {
+            res.status(400).json({errors});
+            return next(errors);
+        }*/
         const deleted =await sheetRepository.deleteSheet(id);
         if (deleted) {
             res.sendStatus(200)
@@ -68,6 +83,8 @@ router.patch('/sheets/:id', async (req, res, next) => {
         const instruments = req.body.instruments;
         const difficulty = req.body.difficulty;
         const imageData = req.body.imageData;
+    
+
 
         const sheet = await sheetRepository.updateSheet(id, title, group_name, instruments, difficulty,imageData);
         if (sheet) {
